Single Player Ping-Pong Game using python & turtle
This game is created in python using turtle. In this game there is one ball and one paddle (stick). Both are turtle objects. Player has to save ball from falling into bottom and other sides (left, right, top) are rigid so ball will bounce back.
Initially Score and High-score are zero and speed of the ball is one. Ball is moving automatically and we have to move paddle (stick) using mouse or keys (left arrow & Right arrow) for saving ball from falling into bottom.
When the score of Player is multiple of 5 (5, 10, 15, …) then the speed of the ball will increase by one. And this feature makes game interesting and challenging.
Once the ball falls into bottom, Score will be reset to zero and speed will be reset to one.